requirements.txt
streamlit
streamlit-webrtc
pytesseract
pdf2image
opencv-python
openpyxl
Pillow
reportlab
